requirements.txt
## 基礎依賴（按順序安裝）
pytz>=2020.1
jsonschema>=3.0
httpx==0.28.1
numpy>=1.26.4,<2.0.0
pandas>=2.2.3

# AI 和 NLP 相關
openai==1.66.3
sentence-transformers==2.5.1
torch==2.6.0
transformers==4.49.0
jieba==0.42.1
tiktoken>=0.5.2  # OpenAI token 計數工具

# Flask 相關套件
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
blinker==1.9.0

# LINE Bot SDK
line-bot-sdk==3.14.5

# 資料庫連接
mysql-connector-python==8.0.26
PyMySQL==1.1.0

# 網路相關
netifaces==0.11.0

# 其他工具
python-dotenv==1.0.0
requests==2.32.3
pickle5>=0.0.11  # Python 3.8+ 的 pickle 擴展

# 翻譯相關
deep-translator==1.11.4

# 視覺化相關
altair>=5.2.0

# 向量處理相關
sentence-transformers==2.5.1
faiss-cpu==1.9.0

# 序列化工具
msgpack>=1.0.5

